# estela-api/docs/api.yaml
swagger: '2.0'
info:
  title: estela API v1.0 Documentation
  description: estela API Swagger Specification
  version: v1
host: 127.0.0.1:8000
schemes:
  - http
basePath: /
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Basic:
    type: basic
security:
  - Basic: []
paths:
...
  /api/projects/{pid}/deploys:
    get:
      operationId: api_projects_deploys_list
      description: ''
      parameters:
        - name: page
          in: query
          description: A page number within the paginated result set.
          required: false
          type: integer
        - name: page_size
          in: query
          description: Number of results to return per page.
          required: false
          type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
              - count
              - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
                x-nullable: true
              previous:
                type: string
                format: uri
                x-nullable: true
              results:
                type: array
                items:
                  $ref: '#/definitions/Deploy'
      tags:
        - api
    post:
      operationId: api_projects_deploys_create
      description: ''
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/DeployCreate'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/DeployCreate'
      tags:
        - api
    parameters:
      - name: pid
        in: path
        required: true
        type: string
...
definitions:
...
  UserDetail:
    description: Username.
    required:
      - username
    type: object
    properties:
      username:
        title: Username
        description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
          only.
        type: string
        pattern: ^[\w.@+-]+$
        maxLength: 150
        minLength: 1
      email:
        title: Email address
        type: string
        format: email
        maxLength: 254
...
  Spider:
    description: Spiders in this deploy.
    required:
      - name
      - project
    type: object
    properties:
      sid:
        title: Sid
        description: A unique integer value identifying this spider.
        type: integer
        readOnly: true
      name:
        title: Name
        description: Spider's name.
        type: string
        maxLength: 1000
        minLength: 1
      project:
        title: Project
        description: Project UUID.
        type: string
        format: uuid
  Deploy:
    required:
      - project
      - user
    type: object
    properties:
      did:
        title: Did
        description: A unique integer value identifying this deploy.
        type: integer
        readOnly: true
      project:
        title: Project
        description: Project UUID.
        type: string
        format: uuid
      user:
        $ref: '#/definitions/UserDetail'
      status:
        title: Status
        description: Deploy status.
        type: string
        enum:
          - SUCCESS
          - BUILDING
          - FAILURE
          - CANCELED
      spiders:
        description: Spiders in this deploy.
        type: array
        items:
          $ref: '#/definitions/Spider'
      created:
        title: Created
        description: Deploy creation date.
        type: string
        format: date-time
        readOnly: true
  DeployCreate:
    type: object
    properties:
      did:
        title: Did
        description: A unique integer value identifying this deploy.
        type: integer
        readOnly: true
      status:
        title: Status
        description: Deploy status.
        type: string
        enum:
          - SUCCESS
          - BUILDING
          - FAILURE
          - CANCELED
      created:
        title: Created
        description: Deploy creation date.
        type: string
        format: date-time
        readOnly: true
      project_zip:
        title: Project zip
        description: Zip file containing the project.
        type: string
        readOnly: true
        format: uri
...